Move pgrx-bench internal runtime API into pgrx module
Create pgrx-bench/src/pgrx.rs and move the pgrx-only public runtime
surface there: the hidden benchmark metadata types, the generated-wrapper
entrypoints, and the module-path enforcement helper.

This keeps the crate root focused on the author-facing API
(Bencher, BatchSize, and black_box) while still preserving the internal
surface area that pgrx-macros and cargo-pgrx need. The new module is
explicitly documented as internal support for generated wrappers and the
bench runner, and the individual items remain #[doc(hidden)] as
requested in review.

The proc-macro expansion paths now point at ::pgrx_bench::pgrx::... so
behavior stays the same even though the code is organized more clearly.

Verification:
- cargo test -p pgrx-bench --features pg16
- cargo check -p pgrx-macros
- cargo check --manifest-path pgrx-examples/benching/Cargo.toml --no-default-features --features 'pg16 pg_bench'
