+++ pgrx-bench/src/lib.rs
@@ -8,12 +8,18 @@
 //LICENSE
 //LICENSE Use of this source code is governed by the MIT license that can be found in the LICENSE file.
 
+pub mod pgrx;
+
 use criterion::{Criterion, measurement::WallTime};
 use oorandom::Rand64;
-use pgrx::JsonB;
-use pgrx::PgTryBuilder;
-use pgrx::pg_sys;
-use serde::{Deserialize, Serialize};
+use ::pgrx::PgTryBuilder;
+use ::pgrx::pg_sys;
+use crate::pgrx::{
+    BenchArtifact, BenchComparison, BenchComparisonEstimate, BenchConfig, BenchDefinition,
+    BenchEstimate, BenchResult, BenchSample, BenchStatus, BenchThroughput, CriterionBenchmark,
+    TransactionMode,
+};
+use serde::Deserialize;
 use serde_json::Value;
 use std::any::Any;
 use std::cell::RefCell;
@@ -22,7 +28,6 @@ use std::fs;
 use std::panic::AssertUnwindSafe;
 use std::path::{Path, PathBuf};
 use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
-use tempfile::TempDir;
 
 /// Re-export of Criterion's `black_box`, which helps keep the optimizer from removing the work
 /// you intend to measure.
@@ -67,187 +72,6 @@ const ARTIFACT_KIND_ESTIMATES_JSON: &str = "criterion_estimates_json";
 const ARTIFACT_KIND_SAMPLE_JSON: &str = "criterion_sample_json";
 const ARTIFACT_KIND_TUKEY_JSON: &str = "criterion_tukey_json";
 const ARTIFACT_KIND_CHANGE_ESTIMATES_JSON: &str = "criterion_change_estimates_json";
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, Eq, PartialEq)]
-#[serde(rename_all = "snake_case")]
-#[doc(hidden)]
-/// Internal transaction metadata generated by `#[pg_bench]`.
-pub enum TransactionMode {
-    Shared,
-    SubtransactionPerBatch,
-    SubtransactionPerIteration,
-}
-
-impl TransactionMode {
-    #[doc(hidden)]
-    /// Returns the SQL/JSON representation used by cargo-pgrx.
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::Shared => "shared",
-            Self::SubtransactionPerBatch => "subtransaction_per_batch",
-            Self::SubtransactionPerIteration => "subtransaction_per_iteration",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal Criterion configuration generated by `#[pg_bench]`.
-pub struct BenchConfig {
-    pub sample_size: usize,
-    pub measurement_time_ms: u64,
-    pub warm_up_time_ms: u64,
-    pub nresamples: usize,
-    pub noise_threshold: f64,
-    pub significance_level: f64,
-}
-
-impl Default for BenchConfig {
-    fn default() -> Self {
-        Self {
-            sample_size: DEFAULT_SAMPLE_SIZE,
-            measurement_time_ms: DEFAULT_MEASUREMENT_TIME_MS,
-            warm_up_time_ms: DEFAULT_WARM_UP_TIME_MS,
-            nresamples: DEFAULT_NRESAMPLES,
-            noise_threshold: DEFAULT_NOISE_THRESHOLD,
-            significance_level: DEFAULT_SIGNIFICANCE_LEVEL,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal benchmark definition generated by `#[pg_bench]`.
-pub struct BenchDefinition {
-    pub schema_name: &'static str,
-    pub bench_name: &'static str,
-    pub function_name: &'static str,
-    pub setup_function: Option<&'static str>,
-    pub transaction_mode: TransactionMode,
-    pub source_file: &'static str,
-    pub source_line: u32,
-    pub config: BenchConfig,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal benchmark result payload returned from the backend to `cargo pgrx bench`.
-pub struct BenchResult {
-    pub schema_name: String,
-    pub bench_name: String,
-    pub function_name: String,
-    pub setup_function: Option<String>,
-    pub transaction_mode: TransactionMode,
-    pub source_file: String,
-    pub source_line: u32,
-    pub criterion_config: BenchConfig,
-    pub status: BenchStatus,
-    pub error_text: Option<String>,
-    pub benchmark: Option<CriterionBenchmark>,
-    pub estimates: Vec<BenchEstimate>,
-    pub samples: Vec<BenchSample>,
-    pub throughput: Option<BenchThroughput>,
-    pub comparison: Option<BenchComparison>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub artifacts: Vec<BenchArtifact>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal benchmark descriptor payload used by `cargo pgrx bench --list`.
-pub struct BenchDescriptor {
-    pub schema_name: String,
-    pub bench_name: String,
-    pub function_name: String,
-    pub setup_function: Option<String>,
-    pub transaction_mode: TransactionMode,
-    pub source_file: String,
-    pub source_line: u32,
-    pub criterion_config: BenchConfig,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
-#[serde(rename_all = "snake_case")]
-#[doc(hidden)]
-/// Internal benchmark execution status.
-pub enum BenchStatus {
-    Ok,
-    Failed,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal copy of Criterion's benchmark identity metadata.
-pub struct CriterionBenchmark {
-    pub group_id: String,
-    pub function_id: Option<String>,
-    pub value_str: Option<String>,
-    pub full_id: String,
-    pub directory_name: String,
-    pub title: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal normalized absolute estimate persisted by the bench runner.
-pub struct BenchEstimate {
-    pub estimate_kind: String,
-    pub point_estimate_ns: f64,
-    pub standard_error_ns: Option<f64>,
-    pub confidence_level: Option<f64>,
-    pub ci_lower_bound_ns: Option<f64>,
-    pub ci_upper_bound_ns: Option<f64>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal normalized sample persisted by the bench runner.
-pub struct BenchSample {
-    pub sample_index: usize,
-    pub iteration_count: u64,
-    pub elapsed_ns: f64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal throughput metadata copied from Criterion output.
-pub struct BenchThroughput {
-    pub kind: String,
-    pub value: f64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal comparison payload derived from Criterion's persisted change analysis.
-pub struct BenchComparison {
-    pub mean: BenchComparisonEstimate,
-    pub median: BenchComparisonEstimate,
-    pub p_value: f64,
-    pub significance_level: f64,
-    pub noise_threshold: f64,
-    pub summary: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal relative estimate produced by Criterion comparison output.
-pub struct BenchComparisonEstimate {
-    pub estimate_kind: String,
-    pub point_estimate: f64,
-    pub standard_error: f64,
-    pub confidence_level: f64,
-    pub ci_lower_bound: f64,
-    pub ci_upper_bound: f64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[doc(hidden)]
-/// Internal persisted Criterion artifact used for storage and baseline replay.
-pub struct BenchArtifact {
-    pub artifact_kind: String,
-    pub media_type: String,
-    pub payload_json: Value,
-}
 
 /// Timing harness passed to a `#[pg_bench]` benchmark function.
 ///
@@ -353,106 +177,6 @@ impl<'a> Bencher<'a> {
                 .to_string()
         })
     }
-}
-
-#[doc(hidden)]
-/// Internal entrypoint used by generated `#[pg_bench]` wrappers.
-pub fn execute_benchmark(
-    definition: BenchDefinition,
-    setup_fn: Option<fn()>,
-    bench_fn: fn(&mut Bencher<'_>),
-    baseline_artifacts: Option<JsonB>,
-) -> JsonB {
-    let definition_for_report = definition.clone();
-    let result = PgTryBuilder::new(AssertUnwindSafe(|| {
-        let baseline_artifacts = baseline_artifacts
-            .map(|baseline_artifacts| {
-                serde_json::from_value::<Vec<BenchArtifact>>(baseline_artifacts.0).map_err(
-                    |error| {
-                        format!("failed to decode persisted Criterion baseline artifacts: {error}")
-                    },
-                )
-            })
-            .transpose()?;
-
-        if let Some(setup_fn) = setup_fn {
-            setup_fn();
-        }
-
-        let mut bencher = Bencher::new(definition.transaction_mode);
-        bench_fn(&mut bencher);
-        let routine = bencher.into_routine()?;
-
-        let tempdir = TempDir::new().map_err(|e| e.to_string())?;
-        // Recreate Criterion's `<bench>/base/*.json` layout up front so the benchmark run and its
-        // comparison analysis are driven by Criterion's own baseline machinery rather than by
-        // cargo-pgrx trying to infer a comparison from normalized tables after the fact.
-        if let Some(baseline_artifacts) = baseline_artifacts.as_deref() {
-            materialize_baseline_artifacts(tempdir.path(), baseline_artifacts)?;
-        }
-
-        let mut criterion =
-            build_criterion(&definition.config, tempdir.path(), baseline_artifacts.is_some());
-        let mut routine = routine;
-
-        criterion.bench_function(definition.bench_name, |criterion_bencher| {
-            run_routine(criterion_bencher, &mut routine, definition.transaction_mode);
-        });
-        criterion.final_summary();
-
-        let report = parse_benchmark_output(
-            definition.clone(),
-            tempdir.path(),
-            baseline_artifacts.as_deref(),
-        )?;
-        Ok::<BenchResult, String>(report)
-    }))
-    .catch_others(|error| Err(caught_error_message(error)))
-    .catch_rust_panic(|error| Err(caught_error_message(error)))
-    .execute();
-
-    match result {
-        Ok(report) => JsonB(serde_json::to_value(report).expect("BenchResult should serialize")),
-        Err(error_text) => JsonB(
-            serde_json::to_value(BenchResult {
-                schema_name: definition_for_report.schema_name.to_string(),
-                bench_name: definition_for_report.bench_name.to_string(),
-                function_name: definition_for_report.function_name.to_string(),
-                setup_function: definition_for_report.setup_function.map(str::to_string),
-                transaction_mode: definition_for_report.transaction_mode,
-                source_file: definition_for_report.source_file.to_string(),
-                source_line: definition_for_report.source_line,
-                criterion_config: definition_for_report.config,
-                status: BenchStatus::Failed,
-                error_text: Some(error_text),
-                benchmark: None,
-                estimates: Vec::new(),
-                samples: Vec::new(),
-                throughput: None,
-                comparison: None,
-                artifacts: Vec::new(),
-            })
-            .expect("BenchResult should serialize"),
-        ),
-    }
-}
-
-#[doc(hidden)]
-/// Internal entrypoint used by generated `#[pg_bench]` descriptor wrappers.
-pub fn describe_benchmark(definition: BenchDefinition) -> JsonB {
-    JsonB(
-        serde_json::to_value(BenchDescriptor {
-            schema_name: definition.schema_name.to_string(),
-            bench_name: definition.bench_name.to_string(),
-            function_name: definition.function_name.to_string(),
-            setup_function: definition.setup_function.map(str::to_string),
-            transaction_mode: definition.transaction_mode,
-            source_file: definition.source_file.to_string(),
-            source_line: definition.source_line,
-            criterion_config: definition.config,
-        })
-        .expect("BenchDescriptor should serialize"),
-    )
 }
 
 fn build_criterion(
@@ -821,11 +545,13 @@ fn parse_throughput(value: Value) -> Option<BenchThroughput> {
     value.as_f64().map(|value| BenchThroughput { kind: kind.to_lowercase(), value })
 }
 
-fn caught_error_message(error: pgrx::pg_sys::panic::CaughtError) -> String {
+fn caught_error_message(error: ::pgrx::pg_sys::panic::CaughtError) -> String {
     match error {
-        pgrx::pg_sys::panic::CaughtError::PostgresError(report)
-        | pgrx::pg_sys::panic::CaughtError::ErrorReport(report) => report.message().to_string(),
-        pgrx::pg_sys::panic::CaughtError::RustPanic { ereport, .. } => {
+        ::pgrx::pg_sys::panic::CaughtError::PostgresError(report)
+        | ::pgrx::pg_sys::panic::CaughtError::ErrorReport(report) => {
+            report.message().to_string()
+        }
+        ::pgrx::pg_sys::panic::CaughtError::RustPanic { ereport, .. } => {
             ereport.message().to_string()
         }
     }
@@ -861,12 +587,6 @@ fn with_internal_subtransaction<R>(f: impl FnOnce() -> R) -> Result<R, String> {
             }
         }
     }
-}
-
-#[doc(hidden)]
-/// Internal helper used by the macro crate to enforce `mod benches`.
-pub const fn module_path_has_benches(module_path: &str) -> bool {
-    ends_with(module_path.as_bytes(), b"::benches") || equals(module_path.as_bytes(), b"benches")
 }
 
 const fn ends_with(value: &[u8], suffix: &[u8]) -> bool {
