+++ pgrx-bench/src/pgrx.rs
@@ -1,0 +1,301 @@
+//! Internal support API for pgrx-generated benchmark wrappers and `cargo pgrx bench`.
+//!
+//! Benchmark authors should use the top-level [`crate::Bencher`], [`crate::BatchSize`], and
+//! [`crate::black_box`] APIs instead of the items in this module.
+
+use ::pgrx::JsonB;
+use ::pgrx::PgTryBuilder;
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+use std::panic::AssertUnwindSafe;
+use tempfile::TempDir;
+
+use crate::Bencher;
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(rename_all = "snake_case")]
+#[doc(hidden)]
+/// Internal transaction metadata generated by `#[pg_bench]`.
+pub enum TransactionMode {
+    Shared,
+    SubtransactionPerBatch,
+    SubtransactionPerIteration,
+}
+
+impl TransactionMode {
+    #[doc(hidden)]
+    /// Returns the SQL/JSON representation used by cargo-pgrx.
+    pub fn as_str(self) -> &'static str {
+        match self {
+            Self::Shared => "shared",
+            Self::SubtransactionPerBatch => "subtransaction_per_batch",
+            Self::SubtransactionPerIteration => "subtransaction_per_iteration",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal Criterion configuration generated by `#[pg_bench]`.
+pub struct BenchConfig {
+    pub sample_size: usize,
+    pub measurement_time_ms: u64,
+    pub warm_up_time_ms: u64,
+    pub nresamples: usize,
+    pub noise_threshold: f64,
+    pub significance_level: f64,
+}
+
+impl Default for BenchConfig {
+    fn default() -> Self {
+        Self {
+            sample_size: crate::DEFAULT_SAMPLE_SIZE,
+            measurement_time_ms: crate::DEFAULT_MEASUREMENT_TIME_MS,
+            warm_up_time_ms: crate::DEFAULT_WARM_UP_TIME_MS,
+            nresamples: crate::DEFAULT_NRESAMPLES,
+            noise_threshold: crate::DEFAULT_NOISE_THRESHOLD,
+            significance_level: crate::DEFAULT_SIGNIFICANCE_LEVEL,
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal benchmark definition generated by `#[pg_bench]`.
+pub struct BenchDefinition {
+    pub schema_name: &'static str,
+    pub bench_name: &'static str,
+    pub function_name: &'static str,
+    pub setup_function: Option<&'static str>,
+    pub transaction_mode: TransactionMode,
+    pub source_file: &'static str,
+    pub source_line: u32,
+    pub config: BenchConfig,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal benchmark result payload returned from the backend to `cargo pgrx bench`.
+pub struct BenchResult {
+    pub schema_name: String,
+    pub bench_name: String,
+    pub function_name: String,
+    pub setup_function: Option<String>,
+    pub transaction_mode: TransactionMode,
+    pub source_file: String,
+    pub source_line: u32,
+    pub criterion_config: BenchConfig,
+    pub status: BenchStatus,
+    pub error_text: Option<String>,
+    pub benchmark: Option<CriterionBenchmark>,
+    pub estimates: Vec<BenchEstimate>,
+    pub samples: Vec<BenchSample>,
+    pub throughput: Option<BenchThroughput>,
+    pub comparison: Option<BenchComparison>,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub artifacts: Vec<BenchArtifact>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal benchmark descriptor payload used by `cargo pgrx bench --list`.
+pub struct BenchDescriptor {
+    pub schema_name: String,
+    pub bench_name: String,
+    pub function_name: String,
+    pub setup_function: Option<String>,
+    pub transaction_mode: TransactionMode,
+    pub source_file: String,
+    pub source_line: u32,
+    pub criterion_config: BenchConfig,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(rename_all = "snake_case")]
+#[doc(hidden)]
+/// Internal benchmark execution status.
+pub enum BenchStatus {
+    Ok,
+    Failed,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal copy of Criterion's benchmark identity metadata.
+pub struct CriterionBenchmark {
+    pub group_id: String,
+    pub function_id: Option<String>,
+    pub value_str: Option<String>,
+    pub full_id: String,
+    pub directory_name: String,
+    pub title: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal normalized absolute estimate persisted by the bench runner.
+pub struct BenchEstimate {
+    pub estimate_kind: String,
+    pub point_estimate_ns: f64,
+    pub standard_error_ns: Option<f64>,
+    pub confidence_level: Option<f64>,
+    pub ci_lower_bound_ns: Option<f64>,
+    pub ci_upper_bound_ns: Option<f64>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal normalized sample persisted by the bench runner.
+pub struct BenchSample {
+    pub sample_index: usize,
+    pub iteration_count: u64,
+    pub elapsed_ns: f64,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal throughput metadata copied from Criterion output.
+pub struct BenchThroughput {
+    pub kind: String,
+    pub value: f64,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal comparison payload derived from Criterion's persisted change analysis.
+pub struct BenchComparison {
+    pub mean: BenchComparisonEstimate,
+    pub median: BenchComparisonEstimate,
+    pub p_value: f64,
+    pub significance_level: f64,
+    pub noise_threshold: f64,
+    pub summary: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal relative estimate produced by Criterion comparison output.
+pub struct BenchComparisonEstimate {
+    pub estimate_kind: String,
+    pub point_estimate: f64,
+    pub standard_error: f64,
+    pub confidence_level: f64,
+    pub ci_lower_bound: f64,
+    pub ci_upper_bound: f64,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[doc(hidden)]
+/// Internal persisted Criterion artifact used for storage and baseline replay.
+pub struct BenchArtifact {
+    pub artifact_kind: String,
+    pub media_type: String,
+    pub payload_json: Value,
+}
+
+#[doc(hidden)]
+/// Internal entrypoint used by generated `#[pg_bench]` wrappers.
+pub fn execute_benchmark(
+    definition: BenchDefinition,
+    setup_fn: Option<fn()>,
+    bench_fn: fn(&mut Bencher<'_>),
+    baseline_artifacts: Option<JsonB>,
+) -> JsonB {
+    let definition_for_report = definition.clone();
+    let result = PgTryBuilder::new(AssertUnwindSafe(|| {
+        let baseline_artifacts = baseline_artifacts
+            .map(|baseline_artifacts| {
+                serde_json::from_value::<Vec<BenchArtifact>>(baseline_artifacts.0).map_err(
+                    |error| {
+                        format!("failed to decode persisted Criterion baseline artifacts: {error}")
+                    },
+                )
+            })
+            .transpose()?;
+
+        if let Some(setup_fn) = setup_fn {
+            setup_fn();
+        }
+
+        let mut bencher = Bencher::new(definition.transaction_mode);
+        bench_fn(&mut bencher);
+        let routine = bencher.into_routine()?;
+
+        let tempdir = TempDir::new().map_err(|e| e.to_string())?;
+        // Recreate Criterion's `<bench>/base/*.json` layout up front so the benchmark run and its
+        // comparison analysis are driven by Criterion's own baseline machinery rather than by
+        // cargo-pgrx trying to infer a comparison from normalized tables after the fact.
+        if let Some(baseline_artifacts) = baseline_artifacts.as_deref() {
+            crate::materialize_baseline_artifacts(tempdir.path(), baseline_artifacts)?;
+        }
+
+        let mut criterion =
+            crate::build_criterion(&definition.config, tempdir.path(), baseline_artifacts.is_some());
+        let mut routine = routine;
+
+        criterion.bench_function(definition.bench_name, |criterion_bencher| {
+            crate::run_routine(criterion_bencher, &mut routine, definition.transaction_mode);
+        });
+        criterion.final_summary();
+
+        let report = crate::parse_benchmark_output(
+            definition.clone(),
+            tempdir.path(),
+            baseline_artifacts.as_deref(),
+        )?;
+        Ok::<BenchResult, String>(report)
+    }))
+    .catch_others(|error| Err(crate::caught_error_message(error)))
+    .catch_rust_panic(|error| Err(crate::caught_error_message(error)))
+    .execute();
+
+    match result {
+        Ok(report) => JsonB(serde_json::to_value(report).expect("BenchResult should serialize")),
+        Err(error_text) => JsonB(
+            serde_json::to_value(BenchResult {
+                schema_name: definition_for_report.schema_name.to_string(),
+                bench_name: definition_for_report.bench_name.to_string(),
+                function_name: definition_for_report.function_name.to_string(),
+                setup_function: definition_for_report.setup_function.map(str::to_string),
+                transaction_mode: definition_for_report.transaction_mode,
+                source_file: definition_for_report.source_file.to_string(),
+                source_line: definition_for_report.source_line,
+                criterion_config: definition_for_report.config,
+                status: BenchStatus::Failed,
+                error_text: Some(error_text),
+                benchmark: None,
+                estimates: Vec::new(),
+                samples: Vec::new(),
+                throughput: None,
+                comparison: None,
+                artifacts: Vec::new(),
+            })
+            .expect("BenchResult should serialize"),
+        ),
+    }
+}
+
+#[doc(hidden)]
+/// Internal entrypoint used by generated `#[pg_bench]` descriptor wrappers.
+pub fn describe_benchmark(definition: BenchDefinition) -> JsonB {
+    JsonB(
+        serde_json::to_value(BenchDescriptor {
+            schema_name: definition.schema_name.to_string(),
+            bench_name: definition.bench_name.to_string(),
+            function_name: definition.function_name.to_string(),
+            setup_function: definition.setup_function.map(str::to_string),
+            transaction_mode: definition.transaction_mode,
+            source_file: definition.source_file.to_string(),
+            source_line: definition.source_line,
+            criterion_config: definition.config,
+        })
+        .expect("BenchDescriptor should serialize"),
+    )
+}
+
+#[doc(hidden)]
+/// Internal helper used by the macro crate to enforce `mod benches`.
+pub const fn module_path_has_benches(module_path: &str) -> bool {
+    crate::ends_with(module_path.as_bytes(), b"::benches")
+        || crate::equals(module_path.as_bytes(), b"benches")
+}
